+++ src/MobileIdAuthenticationHashToSign.php
@@ -115,52 +115,3 @@ class MobileIdAuthenticationHashToSign
     }
 
 }
-
-class MobileIdAuthenticationHashToSignBuilder
-{
-
-    /** @var HashType $hashType */
-    private $hashType;
-
-    /** @var string $hash */
-    private $hash;
-
-    public function getHashType() : ?HashType
-    {
-        return $this->hashType;
-    }
-
-    public function getHash() : ?string
-    {
-        return $this->hash;
-    }
-
-    public function withHashType(string $hashType) : MobileIdAuthenticationHashToSignBuilder
-    {
-        $this->hashType = MobileIdAuthenticationHashToSign::strToHashType($hashType);
-
-        return $this;
-    }
-
-    public function withHashInBase64(string $hash) : MobileIdAuthenticationHashToSignBuilder
-    {
-        $this->hash = $hash;
-        return $this;
-    }
-
-
-    function validateFields() : void
-    {
-        if (is_null($this->getHashType()))
-        {
-            throw new MissingOrInvalidParameterException("Missing hash type");
-        }
-
-    }
-
-    public function build() : MobileIdAuthenticationHashToSign
-    {
-        $this->validateFields();
-        return new MobileIdAuthenticationHashToSign($this);
-    }
-}
