+++ src/MobileIdSignature.php
@@ -26,8 +26,6 @@
  */
 namespace Sk\Mid;
 
-use Sk\Mid\Exception\MissingOrInvalidParameterException;
-
 class MobileIdSignature
 {
     /** @var string $valueInBase64 */
@@ -64,48 +62,4 @@ class MobileIdSignature
 
 }
 
-class MobileIdSignatureBuilder
-{
-    /** @var string $valueInBase64 */
-    private $valueInBase64;
 
-    /** @var string $algorithmName */
-    private $algorithmName;
-
-    public function __construct()
-    {
-    }
-
-    public function getValueInBase64() : string
-    {
-        return $this->valueInBase64;
-    }
-
-    public function getAlgorithmName() : string
-    {
-        return $this->algorithmName;
-    }
-
-    public function withValueInBase64(string $valueInBase64) : MobileIdSignatureBuilder
-    {
-        if (FALSE === base64_decode($valueInBase64))
-        {
-            throw new MissingOrInvalidParameterException("Failed to parse signature value. Input is not valid Base64 string: '" . $valueInBase64 . "'");
-        }
-
-        $this->valueInBase64 = $valueInBase64;
-        return $this;
-    }
-
-    public function withAlgorithmName(string $algorithmName) : MobileIdSignatureBuilder
-    {
-        $this->algorithmName = $algorithmName;
-        return $this;
-    }
-
-    public function build() : MobileIdSignature
-    {
-        return new MobileIdSignature($this);
-    }
-
-}
