+++ src/Rest/SessionStatusPoller.php
@@ -174,63 +174,3 @@ class SessionStatusPoller
     }
 
 }
-class SessionStatusPollerBuilder
-{
-
-    private $connector;
-    /** @var int $pollingSleepTimeoutSeconds */
-    private $pollingSleepTimeoutSeconds = 0;
-    /** @var int $longPollingTimeoutSeconds */
-    private $longPollingTimeoutSeconds = 0;
-
-    /**
-     * @return mixed
-     */
-    public function getConnector()
-    {
-        return $this->connector;
-    }
-
-    /**
-     * @return int
-     */
-    public function getPollingSleepTimeoutSeconds(): int
-    {
-        return $this->pollingSleepTimeoutSeconds;
-    }
-
-    /**
-     * @return int
-     */
-    public function getLongPollingTimeoutSeconds(): int
-    {
-        return $this->longPollingTimeoutSeconds;
-    }
-
-
-    public function withConnector(MobileIdConnector $connector) : SessionStatusPollerBuilder
-    {
-        $this->connector = $connector;
-        return $this;
-    }
-
-    public function withPollingSleepTimeoutSeconds(int $pollingSleepTimeoutSeconds) : SessionStatusPollerBuilder
-    {
-        $this->pollingSleepTimeoutSeconds = $pollingSleepTimeoutSeconds;
-        return $this;
-    }
-
-    public function withLongPollingTimeoutSeconds(int $longPollingTimeoutSeconds) : SessionStatusPollerBuilder
-    {
-        $this->longPollingTimeoutSeconds = $longPollingTimeoutSeconds;
-        return $this;
-    }
-
-    public function build() : SessionStatusPoller
-    {
-        return new SessionStatusPoller($this);
-    }
-
-
-
-}
